split autocomplete out into own file + handles objects
