+++ base/components/PropControls.jsx
@@ -1,0 +1,12 @@
+/**
+ * Import the standard propcontrols, so that they get registered.
+ * 
+ * Usage: import this once in app.jsx
+ * 
+ */
+import PropControlAutoComplete from './PropControlAutoComplete';
+import PropControlPills from './PropControlPills';
+// import SelectDataItemPropControl from './SelectDataItemPropControl'
+
+// no real export - use via PropControl
+export default {};
